Use output flag in publish
